[ADDITION: Message, Termii] - Abstract Termii message added - Dynamic call added for getting api handlers of Base termii client - Added Testing dependencies, and corresponding laravel component
