+++ composer.json
@@ -16,13 +16,22 @@
         }
     ],
 	"require": {
-		"php": ">=7.1",
-        "mane-olawale/termii": "^1.0"
+		"php": ">=7.2",
+        "mane-olawale/termii": "^1.0",
+        "illuminate/notifications": "~5.8.0|^6.0|^7.0|^8.0"
 	},
+    "require-dev": {
+        "php": ">=7.2.0",
+        "phpunit/phpunit": "~8.0|~9.0",
+        "mockery/mockery": "~1.4"
+    },
     "autoload": {
         "psr-4": {
             "ManeOlawale\\Laravel\\Termii\\": "src/"
         }
+    },
+    "scripts": {
+        "tests": "./vendor/bin/phpunit --verbose"
     },
     "extra": {
         "laravel": {
